requirements/base.txt
asgiref==3.8.1
Django==5.1.5
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-filter==23.2
django-environ==0.12.0
django-jazzmin==2.6.0
django-js-asset==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.7
pytz==2024.2
PyYAML==6.0.2
setuptools==75.8.0
sqlparse==0.5.3
uritemplate==4.1.1
